refactor: move invoice settings into folder
